Add: 회원가입 모달 창 분리 후 기능 다시 추가
