+++ src/components/Login/index.jsx
@@ -1,10 +1,9 @@
 import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import LoginModal from '../LoginModal';
+import Join from '../Join';
 
 const Login = ({ isLoginModal, setIsLoginModal }) => {
   const [isJoinModal, setIsJoinModal] = useState(false);
-  const navigate = useNavigate();
 
   return (
     <>
@@ -22,21 +21,7 @@ const Login = ({ isLoginModal, setIsLoginModal }) => {
           }}
         />
       )}
-      {isJoinModal && (
-        <LoginModal
-          title='회원가입'
-          message='계정이 이미 있으신가요?'
-          link='로그인'
-          onClose={() => {
-            setIsJoinModal(false);
-          }}
-          onChange={() => {
-            setIsJoinModal(false);
-            setIsLoginModal(true);
-          }}
-          onClick={() => navigate('/register')}
-        />
-      )}
+      {isJoinModal && <Join setIsLoginModal={setIsLoginModal} setIsJoinModal={setIsJoinModal} />}
     </>
   );
 };
